Add hash.c and dry up the main function.
hash.c is a simple hash table implementation.

added a function called map_dir that will run a function over all
the paths in a directory and add the results to a hash_table.

NOTE: this doesn't work right now, strings are mutatible in C and
it's doing odd things. Don't really know how to fix it right now.